+++ src/lib/dataProcessor/__spec__/mock.ts
@@ -1,6 +1,6 @@
 /* tslint:disable */
 
-import { IJsonRawReport } from "../iJsonReport";
+import { IJsonRawReport } from '@/types';
 
 export const sites = [
   {
